feat: データベース設定の必須化と依存関係の更新
- `dlq-manager`、`event-monitor`、`featureflag`、`policy`、`quota`、`rule-engine`、`saga`、`session`、`vault`、`workflow` などのサービスで、データベース設定がない場合にエラーを発生させるように変更しました。これにより、開発環境やテスト環境以外では必ずデータベースが必要となります。
- `Cargo.toml` ファイルで `rdkafka` のバージョンを `0.37` に更新しました。
- `graphql-gateway` 内のストリーム関連のメソッドで、エラー処理を強化し、接続失敗時に `anyhow::Error` を返すように変更しました。
- `JwksVerifier`、`HealthCollector`、`KeycloakAdminClient`、`HttpSagaClient` の生成メソッドを修正し、エラー処理を追加しました。
- `dynamic_record_repo_impl.rs` でカラムの存在確認を強化し、エラーを適切に伝播するようにしました。

+++ regions/system/server/rust/graphql-gateway/src/infrastructure/grpc/vault_client.rs
@@ -94,11 +94,11 @@ impl VaultGrpcClient {
         offset: Option<i32>,
         limit: Option<i32>,
     ) -> anyhow::Result<Vec<VaultAuditLogEntry>> {
-        // offset/limit は Pagination サブメッセージに移行済み
+        // offset/limit は Pagination サブメッセージに移行済み（Option<i32> 型のため as i32 キャスト不要）
         let request = tonic::Request::new(proto::k1s0::system::vault::v1::ListAuditLogsRequest {
             pagination: Some(proto::k1s0::system::common::v1::Pagination {
-                page: offset.unwrap_or(0) as i32 + 1,
-                page_size: limit.unwrap_or(50) as i32,
+                page: offset.unwrap_or(0) + 1,
+                page_size: limit.unwrap_or(50),
             }),
         });
